requirements.txt
amqp==5.0.6
appdirs==1.4.3
asgiref==3.3.4
billiard==3.6.4.0
black==20.8b1
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.1
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
heroku==0.1.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.1
pep517==0.8.2
Pillow==8.2.0
pkg-resources==0.0.0
progress==1.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
PyJWT==2.1.0
pyparsing==2.4.6
python-dateutil==1.5
pytoml==0.1.21
pytz==2021.1
regex==2021.3.17
requests==2.22.0
retrying==1.3.3
sentry-sdk==1.0.0
six==1.14.0
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.25.8
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
